requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
dnspython==2.7.0
fastapi==0.115.7
h11==0.14.0
idna==3.10
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.10.1
python-dotenv==1.0.1
scalar_fastapi==1.0.3
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
